requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
celery==3.1.20
certifi==2015.11.20.1
cycler==0.9.0
Django==1.8.8
django-autoslug==1.9.3
django-celery==3.1.17
django-cities-light==3.2.0
django-debug-toolbar==1.4
django-grappelli==2.7.3
django-ipware==1.1.3
django-password-reset==0.8.2
django-simple-captcha==0.4.6
djangorestframework==3.3.0
futures==3.0.4
kombu==3.0.33
matplotlib==1.5.0
numpy==1.10.1
pillow==3.0.0
progressbar2==3.6.0
psycopg2==2.6.1
pygeoip==0.3.2
pyparsing==2.1.0
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
sqlparse==0.1.18
Unidecode==0.4.19
